LLM-LAW/LawSummarizer/requirements.txt
python-docx>=1.1.0
llama-cpp-python>=0.2.80